CMS-577: Use context on winter preveiw
